ref: Break circular dependencies between internal files
