+++ packages/react-query/src/interop.ts
@@ -5,7 +5,7 @@ import { CreateTRPCReactOptions } from './shared';
 import {
   CreateReactQueryHooks,
   createHooksInternal,
-} from './shared/hooks/createHooksInternal';
+} from './shared/hooks/createRootHooks';
 
 /**
  * @deprecated use `createTRPCReact` instead
