+++ packages/server/src/deprecated/internals/procedure.ts
@@ -1,8 +1,5 @@
-import { TRPCError } from '../../error/TRPCError';
-import {
-  getCauseFromUnknown,
-  getTRPCErrorFromUnknown,
-} from '../../error/utils';
+import { TRPCError, getTRPCErrorFromUnknown } from '../../error/TRPCError';
+import { getCauseFromUnknown } from '../../error/utils';
 import { InferLast } from '../../types';
 import { ProcedureType } from '../router';
 import {
