+++ packages/server/src/error/TRPCError.ts
@@ -4,6 +4,26 @@ import {
 } from '../error/utils';
 import { TRPC_ERROR_CODE_KEY } from '../rpc/codes';
 
+export function getTRPCErrorFromUnknown(cause: unknown): TRPCError {
+  const error = getErrorFromUnknown(cause);
+  // this should ideally be an `instanceof TRPCError` but for some reason that isn't working
+  // ref https://github.com/trpc/trpc/issues/331
+  if (error.name === 'TRPCError') {
+    return cause as TRPCError;
+  }
+
+  const trpcError = new TRPCError({
+    code: 'INTERNAL_SERVER_ERROR',
+    cause: error,
+    message: error.message,
+  });
+
+  // Inherit stack from error
+  trpcError.stack = error.stack;
+
+  return trpcError;
+}
+
 export class TRPCError extends Error {
   public readonly cause?;
   public readonly code;
