+++ packages/server/src/error/utils.ts
@@ -1,5 +1,3 @@
-import { TRPCError } from './TRPCError';
-
 export function getMessageFromUnknownError(
   err: unknown,
   fallback: string,
@@ -7,7 +5,6 @@ export function getMessageFromUnknownError(
   if (typeof err === 'string') {
     return err;
   }
-
   if (err instanceof Error && typeof err.message === 'string') {
     return err.message;
   }
@@ -20,26 +17,6 @@ export function getErrorFromUnknown(cause: unknown): Error {
   }
   const message = getMessageFromUnknownError(cause, 'Unknown error');
   return new Error(message);
-}
-
-export function getTRPCErrorFromUnknown(cause: unknown): TRPCError {
-  const error = getErrorFromUnknown(cause);
-  // this should ideally be an `instanceof TRPCError` but for some reason that isn't working
-  // ref https://github.com/trpc/trpc/issues/331
-  if (error.name === 'TRPCError') {
-    return cause as TRPCError;
-  }
-
-  const trpcError = new TRPCError({
-    code: 'INTERNAL_SERVER_ERROR',
-    cause: error,
-    message: error.message,
-  });
-
-  // Inherit stack from error
-  trpcError.stack = error.stack;
-
-  return trpcError;
 }
 
 export function getCauseFromUnknown(cause: unknown) {
